requirements.txt
attrs==23.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==2.1.1
colorama==0.4.6
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
numpy==1.25.0
opencv-python==4.7.0.72
packaging==23.1
pinscrape==3.0.3
pluggy==1.2.0
pydotmap==0.1.3
pytest==7.2.0
python-dotenv==1.0.0
requests==2.28.1
soupsieve==2.4.1
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.16
vk-api==11.9.9
